fix build with covfie v0.13.0

+++ src/celeritas/field/detail/CartMapFieldParams.covfie.cc
@@ -66,12 +66,12 @@ struct CartMapFieldParams::Impl
             }
 
             auto affine_translate = covfie::algebra::affine<3>::translation(
-                -inp.min_x, -inp.min_y, -inp.min_z);
+                static_cast<float>(-inp.min_x), static_cast<float>(-inp.min_y), static_cast<float>(-inp.min_z));
 
             auto affine_scale = covfie::algebra::affine<3>::scaling(
-                (inp.num_x - 1) / (inp.max_x - inp.min_x),
-                (inp.num_y - 1) / (inp.max_y - inp.min_y),
-                (inp.num_z - 1) / (inp.max_z - inp.min_z));
+                static_cast<float>((inp.num_x - 1) / (inp.max_x - inp.min_x)),
+                static_cast<float>((inp.num_y - 1) / (inp.max_y - inp.min_y)),
+                static_cast<float>((inp.num_z - 1) / (inp.max_z - inp.min_z)));
 
             using field_t = CovfieFieldTrait<MemSpace::host>::field_t;
             host.field = std::make_unique<field_t>(covfie::make_parameter_pack(
